[challenge repository] add get git repository by id func | move grcmodels gitrepository and gitrepositoryscanresult to grmodels in grlib
