+++ grchallengeapi/domains/challenge/models/data_git_repository_scan_result_filter.go
@@ -1,11 +1,11 @@
 package challengemodels
 
 import (
-	grcenums "github.com/kazekim/backend-engineer-challenge/grchallengeapi/enums"
+	"github.com/kazekim/backend-engineer-challenge/grlib/grenums"
 )
 
 type GitRepositoryScanResultFilterData struct {
 	Id           *string              `json:"id"`
 	RepositoryId *string              `json:"repository_id"`
-	Status       *grcenums.ScanStatus `json:"status"`
+	Status       *grenums.ScanStatus `json:"status"`
 }
