+++ grchallengeapi/domains/challenge/models/response_front_get_git_repository_by_id.go
@@ -1,7 +1,9 @@
 package challengemodels
 
-import grcmodels "github.com/kazekim/backend-engineer-challenge/grchallengeapi/models"
+import (
+	"github.com/kazekim/backend-engineer-challenge/grlib/grmodels"
+)
 
 type FrontGetGitRepositoryByIdResponse struct {
-	Data grcmodels.GitRepository `json:"data"`
+	Data grmodels.GitRepository `json:"data"`
 }
