+++ grchallengeapi/domains/challenge/repositories/repository.go
@@ -2,11 +2,12 @@ package challengerepositories
 
 import (
 	challengemodels "github.com/kazekim/backend-engineer-challenge/grchallengeapi/domains/challenge/models"
-	grcmodels "github.com/kazekim/backend-engineer-challenge/grchallengeapi/models"
 	"github.com/kazekim/backend-engineer-challenge/grlib/grerrors"
+	"github.com/kazekim/backend-engineer-challenge/grlib/grmodels"
 )
 
 type Repository interface {
-	CreateGitRepository(data challengemodels.CreateGitRepositoryData) (*grcmodels.GitRepository, grerrors.Error)
+	CreateGitRepository(data challengemodels.CreateGitRepositoryData) (*grmodels.GitRepository, grerrors.Error)
 	DeleteGitRepositoryById(id string) grerrors.Error
+	GetGitRepositoryById(id string) (*grmodels.GitRepository, grerrors.Error)
 }
