+++ grlib/grmodels/git_repository_scan_result.go
@@ -1,4 +1,4 @@
-package grcmodels
+package grmodels
 
 import (
 	"database/sql"
